requirements.txt
# Core PyQt5
PyQt5==5.15.11
PyQtWebEngine==5.15.7

# Security
cryptography>=41.0.0
bcrypt>=4.0.1
keyring>=24.2.0

# AI Integration
openai>=1.3.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.1

# News & Media
feedparser>=6.0.10
newsapi-python>=0.2.7
python-vlc>=3.0.18122

# Email
imaplib2>=3.6
email-validator>=2.1.0

# Data Processing
pandas>=2.1.3
numpy>=1.24.3
python-dateutil>=2.8.2

# Weather
pyowm>=3.3.0

# Image Processing
Pillow>=10.1.0

# Testing
pytest>=7.4.3
pytest-qt>=4.2.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
apscheduler>=3.10.4
watchdog>=3.0.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0

# Screen capture and AI control
mss>=9.0.1
pyautogui>=0.9.54
opencv-python>=4.8.0
pytesseract>=0.3.10

# For Windows screen capture enhancements
pywin32>=305; sys_platform == 'win32'